move estimates to pages
