+++ src/App.js
@@ -5,8 +5,7 @@ import { auth0Config } from './utils/auth0Config';
 import './App.css';
 import Navigation from './components/Navigation';
 import { InvoicesDash, InvoicesGen } from './pages/Invoices/index';
-import EstimatesGen from './components/Estimates/EstimatesGen';
-import EstimatesDash from './components/Estimates/EstimatesDash';
+import { EstimatesDash, EstimatesGen } from './pages/Estimates/index';
 import { ClientsDash, ClientsGen } from './pages/Clients/index';
 import { BusinessesDash, BusinessesGen } from './pages/Businesses/index';
 import Settings from './components/Settings/Settings';
