+++ src/pages/Estimates/index.js
@@ -1,3 +1,2 @@
 export { default as EstimatesDash } from './EstimatesDash';
 export { default as EstimatesGen } from './EstimatesGen';
-export { default as EstimatesTemplate } from './EstimatesTemplate';
